componentと画面でスクショテスト分割
