+++ AndroidApp/.idea/kotlinc.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="KotlinJpsPluginSettings">
-    <option name="version" value="2.2.21" />
+    <option name="version" value="2.0.21" />
   </component>
 </project>
